move post translation tests into a separate file
